fix cache where no memcache

+++ app/lib/cachefactory.php
@@ -20,33 +20,44 @@ class Cachefactory {
 
 	public function get($id)
 	{
+		if (!$this->cache) return false;
+		
 		$data = $this->cache->get($id);
-
 		return (is_array($data)) ? $data[0] : FALSE;
 	}
 
 	public function save($id, $data, $ttl = 0)
 	{
+		if (!$this->cache) return false;
+		
 		return $this->cache->set($id, array($data, time(), $ttl), $ttl);
 	}
 
 	public function delete($id)
 	{
+		if (!$this->cache) return false;
+		
 		return $this->cache->delete($id);
 	}
 
 	public function clean()
 	{
+		if (!$this->cache) return false;
+		
 		return $this->cache->flush();
 	}
 
 	public function cache_info($type = NULL)
 	{
+		if (!$this->cache) return false;
+		
 		return $this->cache->getStats();
 	}
 
 	public function get_metadata($id)
 	{
+		if (!$this->cache) return false;
+		
 		$stored = $this->cache->get($id);
 
 		if (count($stored) !== 3)
@@ -65,6 +76,8 @@ class Cachefactory {
 	
 	public function connect($host , $port)
 	{ 
+		if (!$this->cache) return false;
+		
 		return $this->cache->addServer($host , $port); 
 	}
